예약옵션 추가(인원수) 및 화면 구성 (#44)
* refactor : queryDsl 사용시 문제가 발생하여 중간 commit

* feat : 날짜별 예약 검색 기능 추가

* comment : page에 대한 주석 추가

* feat : option에 인원수 추가

* refactor : reservationController 리팩토링

* feat : 예약페이지 화면 구성

* feat : 예약 관련 페이지 화면 구성 및 테스트 수정

+++ src/main/java/com/ll/tourdemonde/reservation/service/ReservationService.java
@@ -1,5 +1,6 @@
 package com.ll.tourdemonde.reservation.service;
 
+import com.ll.tourdemonde.global.exception.GlobalException;
 import com.ll.tourdemonde.global.rq.Rq;
 import com.ll.tourdemonde.global.rsData.RsData;
 import com.ll.tourdemonde.global.util.Ut;
@@ -12,9 +13,16 @@ import com.ll.tourdemonde.reservation.entity.Reservation;
 import com.ll.tourdemonde.reservation.entity.ReservationOption;
 import com.ll.tourdemonde.reservation.repository.ReservationRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -26,35 +34,53 @@ public class ReservationService {
     private final MemberService memberService;
     private final Rq rq;
 
+    // 새로운 예약 생성
     @Transactional
-    public Reservation createNewReservation(Place place, ReservationCreateForm form) {
+    public RsData<Reservation> createNewReservation(Place place, ReservationCreateForm form) {
+        try{
         Member seller = memberService.findByUsername(form.getSeller())
                 .orElseThrow(()-> new IllegalArgumentException("해당 유저를 찾을 수 없습니다."));
+
         Reservation reservation = Reservation.builder()
                 .place(place)
                 .seller(seller)
                 .type(form.getType())
                 .build();
-        return reservationRepository.save(reservation);
+        Reservation data = reservationRepository.save(reservation);
+        return new RsData<>("S-success", "예약 생성 성공", data);
+        } catch (Exception e){
+            throw new GlobalException("F-fail", "예약을 생성하지 못했습니다.");
+        }
     }
 
+    // Id로 예약 찾기
     public Reservation findById(Long id) {
         return reservationRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("예약을 찾을 수 없습니다."));
     }
 
+    // 새로운 예약 옵션 생성
     @Transactional
-    public Reservation createNewReservationOption(ReservationOptionForm form) {
-        Reservation reservation = findById(form.getReservationId());
+    public RsData<Reservation> createNewReservationOption(Long reservationId, ReservationOptionForm form) {
+        try {
+            Reservation reservation = findById(reservationId);
 
-        // endDate가 없으면 시작일로 설정
-        form.initEndDateIfNotExists();
+            checkAuthoritiesOfReservation(reservation);
 
-        reservation.addOption(form);
-        reservationRepository.save(reservation);
-        return reservation;
+            // endDate가 없으면 시작일로 설정
+            form.initEndDateIfNotExists();
+
+            reservation.addOption(form);
+            Reservation reservation1 = reservationRepository.save(reservation);
+
+            return new RsData<>("S-success", "예약옵션 생성 성공", reservation1);
+        } catch(Exception e){
+            throw new GlobalException("F-fail", "예약옵션 생성 실패");
+        }
     }
 
+
+    // 장소에 포함된 모든 예약 찾기
     public RsData<List<Reservation>> findAllByPlace(Place place) {
         List<Reservation> reservationList = reservationRepository.findAllByPlaceOrderByIdAsc(place);
 
@@ -66,20 +92,18 @@ public class ReservationService {
         return new RsData<>("S-searchList", "성공", reservationList);
     }
 
+    // 예약 수정
     @Transactional
     public RsData<Reservation> modifyReservation(Reservation reservation, ReservationCreateForm form) {
         // 셀러와 수정한 사람이 동일인인지 확인
-        Member seller = memberService.findByUsername(form.getSeller())
-                .orElseThrow(()-> new IllegalArgumentException("해당 유저를 찾을 수 없습니다."));
-        if (!reservation.getSeller().equals(seller)) {
-            return new RsData<>("f-modify", "권한이 없는 사용자입니다.", null);
-        }
+        checkAuthoritiesOfReservation(reservation);
 
         reservation.setType(form.getType());
 
         return new RsData<>("S-modify", "수정 성공", reservation);
     }
 
+    //예약 옵션 찾기
     public ReservationOption findOptionById(Long reservationId, Long optionId) {
         return reservationRepository.findById(reservationId)
                 .orElseThrow(() -> new IllegalArgumentException("예약을 찾을 수 없습니다."))
@@ -91,13 +115,18 @@ public class ReservationService {
 
     }
 
+    //예약 옵션 수정
     @Transactional
     public ReservationOption modifyReservationOption(Long reservationId, Long optionId, ReservationOptionForm form) {
+        Reservation reservation = reservationRepository.findById(reservationId)
+                .orElseThrow(() -> new IllegalArgumentException("잘못된 입력입니다."));
+
+        checkAuthoritiesOfReservation(reservation);
+
         // endDate 초기화
         form.initEndDateIfNotExists();
 
-        return reservationRepository.findById(reservationId)
-                .orElseThrow(() -> new IllegalArgumentException("잘못된 입력입니다."))
+        return reservation
                 .getOptions().stream()
                 .filter(opt -> opt.getId().equals(optionId))
                 .findFirst()
@@ -105,22 +134,54 @@ public class ReservationService {
                 .modifyValues(
                         Ut.stringToLocalDateTime(form.getStartDate()),
                         Ut.stringToLocalDateTime(form.getEndDate()),
-                        form.getTime(),
-                        form.getPrice()
+                        LocalTime.parse(form.getTime()),
+                        form.getAdultPrice(),
+                        form.getChildrenPrice()
                 );
+
     }
 
+    // 예약 삭제
     @Transactional
     public void deleteReservation(Long reservationId) {
         Reservation reservation = reservationRepository.findById(reservationId)
                 .orElseThrow(() -> new IllegalArgumentException("예약을 찾을 수 없습니다."));
+        checkAuthoritiesOfReservation(reservation);
 
         reservationRepository.delete(reservation);
     }
 
+    // 예약 옵션 삭제
     @Transactional
     public void deleteOption(Reservation reservation, Long optionId) {
+        checkAuthoritiesOfReservation(reservation);
+
         reservation.removeOption(optionId);
     }
 
+    public RsData<Page<ReservationOption>> findByDates(int page, Long placeId, LocalDateTime startDate, LocalDateTime endDate) {
+        // pageable
+        int pageSize = 10;
+        List<Sort.Order> sorts = new ArrayList<>();
+        sorts.add(Sort.Order.desc("startDate"));
+        sorts.add(Sort.Order.desc("endDate"));
+        Pageable pageable = PageRequest.of(page, pageSize, Sort.by(sorts));
+
+        Page<ReservationOption> optionPage =
+                reservationRepository.search(placeId, startDate, endDate, pageable);
+
+        if(optionPage.isEmpty()){
+            return new RsData<>("S-","없음", null);
+        }
+
+        return new RsData<>("S-", "성공", optionPage);
+    }
+
+    private void checkAuthoritiesOfReservation(Reservation reservation) throws GlobalException{
+        Member modifier = rq.getMember();
+
+        if (!reservation.getSeller().equals(modifier) && !modifier.isAdmin()) {
+            throw new GlobalException("F-NoAuthentication", "권한이 없는 사용자입니다.");
+        }
+    }
 }
